pyproject.toml
[project]
name = "wikipedia-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.71.0",
    "fastmcp>=2.12.5",
    "wikipedia>=1.4.0",
]
